[HIB-121] Moving submodules into top level hyp3_process
Had to fix imports in test and modules to support this
