+++ processes/rtc_snap/src/hyp3_process/products/products.py
@@ -4,7 +4,7 @@ from typing import List
 
 import boto3
 
-from outputs import ProcessOutputs
+from ..outputs import ProcessOutputs
 
 s3_resource = boto3.resource('s3')
 s3_client = boto3.client('s3')
